Minor changes to miscellaneous files
Adds the CODEOWNERS file, which dictates what files are owned by whom.
Renames COPYING to COPYING.text (make the file names have the same format).
Removes the alt text for the GPL logo in the README.markdown (was not needed).
Adds a notice about the copyright notice year range in the README.markdown (recommended in the instructions for applying the license).
Updates Gradle to 7.4.2.